+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Current
 
 - Added RequireJS/AMD helpers and documentation
 - Added Python 3 support (thanks to Travis Jensen)
+- Support unnamed URLs resolution.
 
 
 0.6.5 (2013-03-13)